Test Case - breaking changes version using different  module files
